[brain tests] Burst named tuple from the main file
